labs/01-agents/02-gather-evidence/requirements.txt
boto3 
defusedxml 
httpx
strands-agents
bedrock_agentcore
paper-qa